% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_useragent.R
\name{filter_useragent}
\alias{filter_useragent}
\title{filter useragents based on criteria}
\usage{
filter_useragent(
  min_obs = 0,
  software_name = NULL,
  software_type = NULL,
  operating_system_name = NULL,
  layout_engine_name = NULL
)
}
\arguments{
\item{min_obs}{minimum number observed for any given UA string, default 5000}

\item{software_name}{e.g. "Chrome", "Safari", "Firefox", "Edge", etc}

\item{software_type}{e.g. "browser", "bot", "application"}

\item{operating_system_name}{e.g. "Windows", "Linux", "Mac OS X", "macOS", etc}

\item{layout_engine_name}{e.g. "Gecko", "Blink", etc}
}
\value{
character vector of user agents
}
\description{
filter useragents based on criteria
}
\examples{
filter_useragent(
  min_obs = 500000,
  software_name = "Chrome",
  operating_system_name = "Linux"
)
}
