% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_useragent.R
\name{random_useragent}
\alias{random_useragent}
\title{randomly choosing an user-agent string, given a set of conditions}
\usage{
random_useragent(
  min_obs = 5000,
  software_name = NULL,
  software_type = "browser",
  operating_system_name = "Windows",
  layout_engine_name = NULL
)
}
\arguments{
\item{min_obs}{minimum number observed for any given UA string, default 5000}

\item{software_name}{e.g. "Chrome", "Safari", "Firefox", "Edge", etc}

\item{software_type}{e.g. "browser", "bot", "application"}

\item{operating_system_name}{e.g. "Windows", "Linux", "Mac OS X", "macOS", etc}

\item{layout_engine_name}{e.g. "Gecko", "Blink", etc}
}
\value{
a random user agent string, given filter conditions
}
\description{
randomly choosing an user-agent string, given a set of conditions
}
\examples{
random_useragent()
}
